backend/requirements.txt
flask
flask-cors
tensorflow-macos
tensorflow-metal
Pillow
numpy
gunicorn

